Added Ai debug options Fixed Transform issues Added Rock model Added Fix Aspect ratio in editor option Fix issues with engine exit Added Vector utils class

+++ GraphicsEngine/Source/Core/Audio/AudioEngine.cpp
@@ -32,8 +32,11 @@ void AudioEngine::Startup()
 
 void AudioEngine::Shutdown()
 {
-	Instance->Terminate();
-	SafeDelete(Instance);
+	if (Instance != nullptr)
+	{
+		Instance->Terminate();
+		SafeDelete(Instance);
+	}
 }
 
 AudioEngine * AudioEngine::Get()
